+++ data/duke.txt
@@ -1,2 +1,2 @@
-D | 0 | eat ryan  | 2023-03-22T09:00
-E | 0 | eat shishir  | 2023-09-22T09:00--2009-09-09T09:00
+T | 0 | eat breakfast | 
+E | 0 | host a party  | 2023-05-22T09:00--2023-05-22T12:00
